+++ src/fusemt.rs
@@ -36,6 +36,26 @@ impl<'a> IntoRequestInfo for fuse::Request<'a> {
     }
 }
 
+fn fuse_fileattr(attr: FileAttr, ino: u64) -> fuse::FileAttr {
+    fuse::FileAttr {
+        ino: ino,
+        size: attr.size,
+        blocks: attr.blocks,
+        atime: attr.atime,
+        mtime: attr.mtime,
+        ctime: attr.ctime,
+        crtime: attr.crtime,
+        kind: attr.kind,
+        perm: attr.perm,
+        nlink: attr.nlink,
+        uid: attr.uid,
+        gid: attr.gid,
+        rdev: attr.rdev,
+        flags: attr.flags,
+    }
+}
+
+#[derive(Debug)]
 pub struct FuseMT<T> {
     target: T,
     inodes: InodeTable,
@@ -117,12 +137,11 @@ impl<T: FilesystemMT + Sync + Send + 'static> fuse::Filesystem for FuseMT<T> {
         let parent_path = get_path!(self, parent, reply);
         debug!("lookup: {:?}, {:?}", parent_path, name);
         let path = Arc::new((*parent_path).clone().join(name));
-        match self.target.lookup(req.info(), Path::new(&*parent_path), name) {
-            Ok((ref ttl, ref mut attr)) => {
+        match self.target.getattr(req.info(), &path, None) {
+            Ok((ttl, attr)) => {
                 let (ino, generation) = self.inodes.add_or_get(path.clone());
                 self.inodes.lookup(ino);
-                attr.ino = ino;
-                reply.entry(ttl, attr, generation);
+                reply.entry(&ttl, &fuse_fileattr(attr, ino), generation);
             },
             Err(e) => reply.error(e),
         }
@@ -140,9 +159,8 @@ impl<T: FilesystemMT + Sync + Send + 'static> fuse::Filesystem for FuseMT<T> {
         let path = get_path!(self, ino, reply);
         debug!("getattr: {:?}", path);
         match self.target.getattr(req.info(), &path, None) {
-            Ok((ref ttl, ref mut attr)) => {
-                attr.ino = ino;
-                reply.attr(ttl, attr)
+            Ok((ttl, attr)) => {
+                reply.attr(&ttl, &fuse_fileattr(attr, ino))
             },
             Err(e) => reply.error(e),
         }
@@ -213,7 +231,7 @@ impl<T: FilesystemMT + Sync + Send + 'static> fuse::Filesystem for FuseMT<T> {
         }
 
         match self.target.getattr(req.info(), &path, fh) {
-            Ok((ref ttl, ref attr)) => reply.attr(ttl, attr),
+            Ok((ttl, attr)) => reply.attr(&ttl, &fuse_fileattr(attr, ino)),
             Err(e) => reply.error(e),
         }
    }
@@ -231,10 +249,9 @@ impl<T: FilesystemMT + Sync + Send + 'static> fuse::Filesystem for FuseMT<T> {
         let parent_path = get_path!(self, parent, reply);
         debug!("mknod: {:?}/{:?}", parent_path, name);
         match self.target.mknod(req.info(), &parent_path, name, mode, rdev) {
-            Ok((ref ttl, ref mut attr)) => {
+            Ok((ttl, attr)) => {
                 let (ino, generation) = self.inodes.add(Arc::new(parent_path.join(name)));
-                attr.ino = ino;
-                reply.entry(ttl, attr, generation)
+                reply.entry(&ttl, &fuse_fileattr(attr, ino), generation)
             },
             Err(e) => reply.error(e),
         }
@@ -244,10 +261,9 @@ impl<T: FilesystemMT + Sync + Send + 'static> fuse::Filesystem for FuseMT<T> {
         let parent_path = get_path!(self, parent, reply);
         debug!("mkdir: {:?}/{:?}", parent_path, name);
         match self.target.mkdir(req.info(), &parent_path, name, mode) {
-            Ok((ref ttl, ref mut attr)) => {
+            Ok((ttl, attr)) => {
                 let (ino, generation) = self.inodes.add(Arc::new(parent_path.join(name)));
-                attr.ino = ino;
-                reply.entry(ttl, attr, generation)
+                reply.entry(&ttl, &fuse_fileattr(attr, ino), generation)
             },
             Err(e) => reply.error(e),
         }
@@ -278,10 +294,9 @@ impl<T: FilesystemMT + Sync + Send + 'static> fuse::Filesystem for FuseMT<T> {
         let parent_path = get_path!(self, parent, reply);
         debug!("symlink: {:?}/{:?} -> {:?}", parent_path, name, link);
         match self.target.symlink(req.info(), &parent_path, name, link) {
-            Ok((ref ttl, ref mut attr)) => {
+            Ok((ttl, attr)) => {
                 let (ino, generation) = self.inodes.add(Arc::new(parent_path.join(name)));
-                attr.ino = ino;
-                reply.entry(ttl, attr, generation)
+                reply.entry(&ttl, &fuse_fileattr(attr, ino), generation)
             },
             Err(e) => reply.error(e),
         }
@@ -305,12 +320,11 @@ impl<T: FilesystemMT + Sync + Send + 'static> fuse::Filesystem for FuseMT<T> {
         let newparent_path = get_path!(self, newparent, reply);
         debug!("link: {:?} -> {:?}/{:?}", path, newparent_path, newname);
         match self.target.link(req.info(), &path, &newparent_path, newname) {
-            Ok((ref ttl, ref mut attr)) => {
+            Ok((ttl, attr)) => {
                 // NOTE: this results in the new link having a different inode from the original.
                 // This is needed because our inode table is a 1:1 map between paths and inodes.
                 let (new_ino, generation) = self.inodes.add(Arc::new(newparent_path.join(newname)));
-                attr.ino = new_ino;
-                reply.entry(ttl, attr, generation);
+                reply.entry(&ttl, &fuse_fileattr(attr, new_ino), generation);
             },
             Err(e) => reply.error(e),
         }
@@ -584,10 +598,10 @@ impl<T: FilesystemMT + Sync + Send + 'static> fuse::Filesystem for FuseMT<T> {
         let parent_path = get_path!(self, parent, reply);
         debug!("create: {:?}/{:?} (mode={:#o}, flags={:#x})", parent_path, name, mode, flags);
         match self.target.create(req.info(), &parent_path, name, mode, flags) {
-            Ok(mut create) => {
+            Ok(create) => {
                 let (ino, generation) = self.inodes.add(Arc::new(parent_path.join(name)));
-                create.attr.ino = ino;
-                reply.created(&create.ttl, &create.attr, generation, create.fh, create.flags);
+                let attr = fuse_fileattr(create.attr, ino);
+                reply.created(&create.ttl, &attr, generation, create.fh, create.flags);
             },
             Err(e) => reply.error(e),
         }